Refactoring the project to get rid of Lab 3 stuff.
