+++ .idea/misc.xml
@@ -3,6 +3,8 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/Users/cup_k/Team-8-TCSS-450/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.403125" />
+        <entry key="..\:/Users/cup_k/Team-8-TCSS-450/app/src/main/res/drawable/ic_launcher_background.xml" value="0.403125" />
         <entry key="..\:/Users/cup_k/Team-8-TCSS-450/app/src/main/res/drawable/ic_weather_black_24.xml" value="0.28020833333333334" />
         <entry key="..\:/Users/cup_k/Team-8-TCSS-450/app/src/main/res/layout/activity_main.xml" value="0.2" />
         <entry key="..\:/Users/cup_k/Team-8-TCSS-450/app/src/main/res/layout/fargment_contacts.xml" value="0.2640625" />
