--- app/src/main/java/edu/uw/tcss450/lab3_authentication/AuthActivity.java
@@ -1,14 +1,0 @@
-package edu.uw.tcss450.lab3_authentication;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class AuthActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_auth);
-    }
-}
