--- app/src/main/java/edu/uw/tcss450/lab3_authentication/ui/HomeFragment.java
@@ -1,44 +1,0 @@
-package edu.uw.tcss450.lab3_authentication.ui;
-
-import android.os.Bundle;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-import androidx.lifecycle.ViewModelProvider;
-
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import edu.uw.tcss450.lab3_authentication.R;
-import edu.uw.tcss450.lab3_authentication.databinding.FragmentHomeBinding;
-import edu.uw.tcss450.lab3_authentication.databinding.FragmentRegisterBinding;
-import edu.uw.tcss450.lab3_authentication.model.UserInfoViewModel;
-
-/**
- * A simple {@link Fragment} subclass.
- */
-public class HomeFragment extends Fragment {
-
-    public HomeFragment() {
-        // Required empty public constructor
-    }
-
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_home, container, false);
-    }
-
-    @Override
-    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
-        UserInfoViewModel model = new ViewModelProvider(getActivity())
-                .get(UserInfoViewModel.class);
-
-        FragmentHomeBinding.bind(getView()).textHello.setText("Welcome!");
-    }
-}
