--- app/src/main/java/edu/uw/tcss450/lab3_authentication/ui/account/AccountViewModel.java
@@ -1,6 +1,0 @@
-package edu.uw.tcss450.lab3_authentication.ui.account;
-
-import androidx.lifecycle.ViewModel;
-
-public class AccountViewModel extends ViewModel {
-}
