--- app/src/main/java/edu/uw/tcss450/lab3_authentication/ui/contacts/ContactsViewModel.java
@@ -1,14 +1,0 @@
-package edu.uw.tcss450.lab3_authentication.ui.contacts;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.fragment.app.Fragment;
-import androidx.lifecycle.ViewModel;
-
-
-public class ContactsViewModel extends ViewModel {
-
-}
