--- app/src/main/java/edu/uw/tcss450/lab3_authentication/ui/messages/MessagesViewModel.java
@@ -1,6 +1,0 @@
-package edu.uw.tcss450.lab3_authentication.ui.messages;
-
-import androidx.lifecycle.ViewModel;
-
-public class MessagesViewModel extends ViewModel {
-}
