--- app/src/main/java/edu/uw/tcss450/lab3_authentication/ui/weather/WeatherFragment.java
@@ -1,24 +1,0 @@
-package edu.uw.tcss450.lab3_authentication.ui.weather;
-
-import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import edu.uw.tcss450.lab3_authentication.R;
-
-/**
- * A simple {@link Fragment} subclass.
- */
-public class WeatherFragment extends Fragment {
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_weather, container, false);
-    }
-}
