--- app/src/main/java/edu/uw/tcss450/lab3_authentication/ui/weather/WeatherViewModel.java
@@ -1,6 +1,0 @@
-package edu.uw.tcss450.lab3_authentication.ui.weather;
-
-import androidx.lifecycle.ViewModel;
-
-public class WeatherViewModel extends ViewModel {
-}
