+++ app/src/main/java/edu/uw/tcss450/messaging_final_project/AuthActivity.java
@@ -1,0 +1,14 @@
+package edu.uw.tcss450.messaging_final_project;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+public class AuthActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_auth);
+    }
+}
