+++ app/src/main/java/edu/uw/tcss450/messaging_final_project/ui/account/AccountViewModel.java
@@ -1,0 +1,6 @@
+package edu.uw.tcss450.messaging_final_project.ui.account;
+
+import androidx.lifecycle.ViewModel;
+
+public class AccountViewModel extends ViewModel {
+}
