+++ app/src/main/java/edu/uw/tcss450/messaging_final_project/ui/auth/register/RegisterFragment.java
@@ -1,0 +1,163 @@
+package edu.uw.tcss450.messaging_final_project.ui.auth.register;
+
+import static edu.uw.tcss450.messaging_final_project.utils.PasswordValidator.*;
+import static edu.uw.tcss450.messaging_final_project.utils.PasswordValidator.checkClientPredicate;
+
+import android.os.Bundle;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.ViewModelProvider;
+import androidx.navigation.Navigation;
+
+import android.util.Log;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import org.json.JSONException;
+import org.json.JSONObject;
+
+import edu.uw.tcss450.messaging_final_project.databinding.FragmentRegisterBinding;
+import edu.uw.tcss450.messaging_final_project.utils.PasswordValidator;
+
+
+/**
+ * A simple {@link Fragment} subclass.
+ */
+public class RegisterFragment extends Fragment {
+
+    private FragmentRegisterBinding binding;
+
+    private RegisterViewModel mRegisterModel;
+
+    private PasswordValidator mNameValidator = checkPwdLength(1);
+
+    private PasswordValidator mEmailValidator = checkPwdLength(2)
+            .and(checkExcludeWhiteSpace())
+            .and(checkPwdSpecialChar("@"));
+
+    private PasswordValidator mPassWordValidator =
+            checkClientPredicate(pwd -> pwd.equals(binding.editPassword2.getText().toString()))
+                    .and(checkPwdLength(7))
+                    .and(checkPwdSpecialChar())
+                    .and(checkExcludeWhiteSpace())
+                    .and(checkPwdDigit())
+                    .and(checkPwdLowerCase().or(checkPwdUpperCase()));
+
+    public RegisterFragment() {
+        // Required empty public constructor
+    }
+
+    @Override
+    public void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        mRegisterModel = new ViewModelProvider(getActivity())
+                .get(RegisterViewModel.class);
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        binding = FragmentRegisterBinding.inflate(inflater);
+        return binding.getRoot();
+    }
+
+    @Override
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+
+        binding.buttonRegister.setOnClickListener(this::attemptRegister);
+        mRegisterModel.addResponseObserver(getViewLifecycleOwner(),
+                this::observeResponse);
+    }
+
+    private void attemptRegister(final View button) {
+        validateFirst();
+    }
+
+    private void validateFirst() {
+        mNameValidator.processResult(
+                mNameValidator.apply(binding.editFirst.getText().toString().trim()),
+                this::validateLast,
+                result -> binding.editFirst.setError("Please enter a first name."));
+    }
+
+    private void validateLast() {
+        mNameValidator.processResult(
+                mNameValidator.apply(binding.editLast.getText().toString().trim()),
+                this::validateEmail,
+                result -> binding.editLast.setError("Please enter a last name."));
+    }
+
+    private void validateEmail() {
+        mEmailValidator.processResult(
+                mEmailValidator.apply(binding.editEmail.getText().toString().trim()),
+                this::validatePasswordsMatch,
+                result -> binding.editEmail.setError("Please enter a valid Email address."));
+    }
+
+    private void validatePasswordsMatch() {
+        PasswordValidator matchValidator =
+                checkClientPredicate(
+                        pwd -> pwd.equals(binding.editPassword2.getText().toString().trim()));
+
+        mEmailValidator.processResult(
+                matchValidator.apply(binding.editPassword1.getText().toString().trim()),
+                this::validatePassword,
+                result -> binding.editPassword1.setError("Passwords must match."));
+    }
+
+    private void validatePassword() {
+        mPassWordValidator.processResult(
+                mPassWordValidator.apply(binding.editPassword1.getText().toString()),
+                this::verifyAuthWithServer,
+                result -> binding.editPassword1.setError("Please enter a valid Password."));
+    }
+
+    private void verifyAuthWithServer() {
+        mRegisterModel.connect(
+                binding.editFirst.getText().toString(),
+                binding.editLast.getText().toString(),
+                binding.editEmail.getText().toString(),
+                binding.editPassword1.getText().toString());
+        //This is an Asynchronous call. No statements after should rely on the
+        // result of connect()
+    }
+
+    private void navigateToLogin() {
+        RegisterFragmentDirections.ActionRegisterFragmentToLoginFragment directions =
+                RegisterFragmentDirections.actionRegisterFragmentToLoginFragment();
+
+        directions.setEmail(binding.editEmail.getText().toString());
+        directions.setPassword(binding.editPassword1.getText().toString());
+
+        Navigation.findNavController(getView()).navigate(directions);
+
+    }
+
+    /**
+     * An observer on the HTTP Response from the web server. This observer should be
+     * attached to SignInViewModel.
+     *
+     * @param response the Response from the server
+     */
+    private void observeResponse(final JSONObject response) {
+        if (response.length() > 0) {
+            if (response.has("code")) {
+                try {
+                    binding.editEmail.setError("Error Authenticating: " + response.getJSONObject("data").getString("message"));
+                } catch (JSONException e) {
+                    Log.e("JSON Parse Error", e.getMessage());
+                }
+            } else {
+                navigateToLogin();
+            }
+        } else {
+            Log.d("JSON Response", "No Response");
+        }
+    }
+
+
+}
