+++ app/src/main/java/edu/uw/tcss450/messaging_final_project/ui/auth/signin/SignInFragment.java
@@ -1,0 +1,141 @@
+package edu.uw.tcss450.messaging_final_project.ui.auth.signin;
+
+import static edu.uw.tcss450.messaging_final_project.utils.PasswordValidator.*;
+
+import android.os.Bundle;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.ViewModelProvider;
+import androidx.navigation.Navigation;
+
+import android.util.Log;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import org.json.JSONException;
+import org.json.JSONObject;
+
+import edu.uw.tcss450.messaging_final_project.databinding.FragmentSignInBinding;
+import edu.uw.tcss450.messaging_final_project.utils.PasswordValidator;
+
+/**
+ * A simple {@link Fragment} subclass.
+ */
+public class SignInFragment extends Fragment {
+
+    private FragmentSignInBinding binding;
+    private SignInViewModel mSignInModel;
+
+    private PasswordValidator mEmailValidator = checkPwdLength(2)
+            .and(checkExcludeWhiteSpace())
+            .and(checkPwdSpecialChar("@"));
+
+    private PasswordValidator mPassWordValidator = checkPwdLength(1)
+            .and(checkExcludeWhiteSpace());
+
+    public SignInFragment() {
+        // Required empty public constructor
+    }
+
+    @Override
+    public void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        mSignInModel = new ViewModelProvider(getActivity())
+                .get(SignInViewModel.class);
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        binding = FragmentSignInBinding.inflate(inflater);
+        // Inflate the layout for this fragment
+        return binding.getRoot();
+    }
+
+    @Override
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+
+        binding.buttonToRegister.setOnClickListener(button ->
+                Navigation.findNavController(getView()).navigate(
+                        SignInFragmentDirections.actionLoginFragmentToRegisterFragment()
+                ));
+
+        binding.buttonSignIn.setOnClickListener(this::attemptSignIn);
+
+        mSignInModel.addResponseObserver(
+                getViewLifecycleOwner(),
+                this::observeResponse);
+
+        SignInFragmentArgs args = SignInFragmentArgs.fromBundle(getArguments());
+        binding.editEmail.setText(args.getEmail().equals("default") ? "" : args.getEmail());
+        binding.editPassword.setText(args.getPassword().equals("default") ? "" : args.getPassword());
+    }
+
+    private void attemptSignIn(final View button) {
+        validateEmail();
+    }
+
+    private void validateEmail() {
+        mEmailValidator.processResult(
+                mEmailValidator.apply(binding.editEmail.getText().toString().trim()),
+                this::validatePassword,
+                result -> binding.editEmail.setError("Please enter a valid Email address."));
+    }
+
+    private void validatePassword() {
+        mPassWordValidator.processResult(
+                mPassWordValidator.apply(binding.editPassword.getText().toString()),
+                this::verifyAuthWithServer,
+                result -> binding.editPassword.setError("Please enter a valid Password."));
+    }
+
+    private void verifyAuthWithServer() {
+        mSignInModel.connect(binding.editEmail.getText().toString(), binding.editPassword.getText().toString());
+        //This is an Asynchronous call. No statements after should rely on the
+        // result of connect().
+
+    }
+
+    /**
+     * Helper to abstract the navigation to the Activity past Authentication.
+     *
+     * @param email users email
+     * @param jwt   the JSON Web Token supplied by the server
+     */
+    private void navigateToSuccess(final String email, final String jwt) {
+        Navigation.findNavController(getView())
+                .navigate(SignInFragmentDirections
+                        .actionLoginFragmentToMainActivity(email, jwt));
+    }
+
+    /**
+     * An observer on the HTTP Response from the web server. This observer should be
+     * attached to SignInViewModel.
+     *
+     * @param response the Response from the server
+     */
+    private void observeResponse(final JSONObject response) {
+        if (response.length() > 0) {
+            if (response.has("code")) {
+                try {
+                    binding.editEmail.setError("Error Authenticating: " + response.getJSONObject("data").getString("message"));
+                } catch (JSONException e) {
+                    Log.e("JSON Parse Error", e.getMessage());
+                }
+            } else {
+                try {
+                    navigateToSuccess(binding.editEmail.getText().toString(), response.getString("token"));
+                } catch (JSONException e) {
+                    Log.e("JSON Parse Error", e.getMessage());
+                }
+            }
+        } else {
+            Log.d("JSON Response", "No Response");
+        }
+
+    }
+}
