+++ app/src/main/java/edu/uw/tcss450/messaging_final_project/ui/contacts/ContactsViewModel.java
@@ -1,0 +1,8 @@
+package edu.uw.tcss450.messaging_final_project.ui.contacts;
+
+import androidx.lifecycle.ViewModel;
+
+
+public class ContactsViewModel extends ViewModel {
+
+}
