+++ app/src/main/java/edu/uw/tcss450/messaging_final_project/ui/messages/MessagesViewModel.java
@@ -1,0 +1,6 @@
+package edu.uw.tcss450.messaging_final_project.ui.messages;
+
+import androidx.lifecycle.ViewModel;
+
+public class MessagesViewModel extends ViewModel {
+}
