+++ app/src/main/java/edu/uw/tcss450/messaging_final_project/ui/weather/WeatherViewModel.java
@@ -1,0 +1,6 @@
+package edu.uw.tcss450.messaging_final_project.ui.weather;
+
+import androidx.lifecycle.ViewModel;
+
+public class WeatherViewModel extends ViewModel {
+}
