+++ app/src/main/res/values/strings.xml
@@ -8,8 +8,6 @@
     <string name="hint_password_again">Retype Password</string>
     <string name="action_sign_in_short">Sign in</string>
     <string name="action_register_short">Register</string>
-    <string name="display_header_title">Phish Shows Welcomes:</string>
-    <string name="image_description_logo">The Phish Logo</string>
 
     <string name="action_signin">Sign In</string>
     <string name="action_next">Next</string>
